Extract SpellRunes from Spell.kt
